Replace float by double in Cython code.

+++ quantlib/instruments/credit_default_swap.pyx
@@ -56,7 +56,7 @@ cdef class CreditDefaultSwap(Instrument):
     """
 
 
-    def __init__(self, int side, float notional, float spread, Schedule schedule,
+    def __init__(self, int side, double notional, double spread, Schedule schedule,
                  int payment_convention,
                  DayCounter day_counter, bool settles_accrual=True,
                  bool pays_at_default_time=True,
